+++ Cortex/MedableCortex.framework/Headers/MDPaginationHelper.h
@@ -107,20 +107,20 @@ NS_ASSUME_NONNULL_BEGIN
  * Convenience Initializer to paginate a context using a different paging field other than `_id`.
  *
  * @param context The context name (plural context) for the objects to be paginated.
- * @param pagingField The name of the field/property to use to paginate the objects. It has to be marked as indexed and unique in Medable API.
+ * @param pagingField The name of the field/property to use to paginate the objects. It has to be marked as indexed and unique in Medable API. If `nil` is passed, skip+limit pagination method is used.
  * @param pageSize The size of the page.
  * @param cacheResults Specifies whether to cache the results or not.
  * @param inverseOrder Pass in `YES` for inverse sorting of objects.
  * @return Initialized instance.
  */
 + (nullable instancetype)paginatorWithContext:(NSString *)context
-                                  pagingField:(NSString *)pagingField
+                                  pagingField:(nullable NSString *)pagingField
                                      pageSize:(NSUInteger)pageSize
                                  cacheResults:(BOOL)cacheResults
                                  inverseOrder:(BOOL)inverseOrder;
 
 /**
- * Convenience initializer to paginate a list in a path.
+ * Convenience initializer to paginate a list in a path with `_id` as the paging field.
  *
  * The field/property used for pagination is `_id`.
  *
@@ -133,6 +133,24 @@ NS_ASSUME_NONNULL_BEGIN
  * @return Initialized instance.
  */
 + (nullable instancetype)paginatorWithPath:(NSString *)path
+                                  pageSize:(NSUInteger)pageSize
+                              cacheResults:(BOOL)cacheResults
+                              inverseOrder:(BOOL)inverseOrder;
+
+/**
+ * Convenience initializer to paginate a list in a path with (optional) `pagingField` as the paging field.
+ *
+ * Use this initializer to paginate list properties with a path: i.e: `GET /path/to/listProperty?where={"pagingField":{"$gt":"lastValue"}}&limit=pageSize`
+ *
+ * @param path The path to the list property. The path ends with the property name. Don't finish the path with a trailing forward slash. ObjectIds might be used in the path and also in array subpathing. i.e.: path/to/listProperty
+ * @param pagingField The name of the field/property to use to paginate the objects. It has to be marked as indexed and unique in Medable API. If `nil` is passed, skip+limit pagination method is used.
+ * @param pageSize The size of the page.
+ * @param cacheResults Specifies whether to cache the results or not.
+ * @param inverseOrder Pass in `YES` for inverse sorting of objects.
+ * @return Initialized instance.
+ */
++ (nullable instancetype)paginatorWithPath:(NSString *)path
+                               pagingField:(nullable NSString *)pagingField
                                   pageSize:(NSUInteger)pageSize
                               cacheResults:(BOOL)cacheResults
                               inverseOrder:(BOOL)inverseOrder;
@@ -218,8 +236,10 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  * Load next page. Results through callback and / or delegate properties.
  * If the paginator is not in Idle state, this operation will be queued.
+ *
+ * @return The request URL. Note: `nil` could be returned if the paginator helper is busy and this call gets delayed and queued.
  */
-- (void)loadNextPage;
+- (nullable NSString *)loadNextPage;
 
 /**
  * Loads all objects. Keeps loading pages until all the objects are loaded. Not recommended.